Fixing circular dependencies in application-system-form
