+++ libs/application/core/src/index.ts
@@ -19,6 +19,7 @@ export * from './lib/ApplicationTemplateHelper'
 export * from './lib/messages'
 export * from './lib/configurationUtils'
 export * from './lib/constants'
+export * from './lib/getValueViaPath'
 
 export * from './validation/validators'
 export * from './validation/AnswerValidator'
